+++ src/compiler_top.rs
@@ -138,8 +138,11 @@ impl Linker {
         );
 
         self.with_file_builder(file_id, |builder| {
-            let _panic_guard =
-                SpanDebugger::new("gather_initial_file_data in add_file", builder.file_data);
+            let _panic_guard = SpanDebugger::new(
+                "gather_initial_file_data in add_file",
+                &builder.file_data.file_identifier,
+                builder.file_data,
+            );
             gather_initial_file_data(builder);
         });
 
@@ -168,8 +171,11 @@ impl Linker {
             file_data.tree = tree;
 
             self.with_file_builder(file_id, |builder| {
-                let _panic_guard =
-                    SpanDebugger::new("gather_initial_file_data in update_file", builder.file_data);
+                let _panic_guard = SpanDebugger::new(
+                    "gather_initial_file_data in update_file",
+                    &builder.file_data.file_identifier,
+                    builder.file_data,
+                );
                 gather_initial_file_data(builder);
             });
 
@@ -209,22 +215,12 @@ impl Linker {
         }
 
         flatten_all_globals(self);
-        if config.debug_print_module_contents {
-            config.for_each_debug_module(&self.modules, |md| {
-                md.print_flattened_module(&self.files[md.link_info.file]);
-            });
-        }
         if config.early_exit == EarlyExitUpTo::Flatten {
             return;
         }
 
         typecheck_all_modules(self);
 
-        if config.debug_print_module_contents {
-            config.for_each_debug_module(&self.modules, |md| {
-                md.print_flattened_module(&self.files[md.link_info.file]);
-            });
-        }
         if config.early_exit == EarlyExitUpTo::AbstractTypecheck {
             return;
         }
@@ -238,12 +234,7 @@ impl Linker {
         // Make an initial instantiation of all modules
         // Won't be possible once we have template modules
         for (id, md) in &self.modules {
-            //md.print_flattened_module();
             // Already instantiate any modules without parameters
-            // Currently this is all modules
-            let span_debug_message = format!("instantiating {}", &md.link_info.name);
-            let _panic_guard =
-                SpanDebugger::new(&span_debug_message, &self.files[md.link_info.file]);
             // Can immediately instantiate modules that have no template args
             if md.link_info.template_parameters.is_empty() {
                 let _inst = md.instantiations.instantiate(