+++ src/flattening/initialization.rs
@@ -275,7 +275,7 @@ impl InitializationContext<'_> {
 pub fn gather_initial_file_data(mut builder: FileBuilder) {
     assert!(builder.file_data.associated_values.is_empty());
 
-    let mut cursor = match Cursor::new_at_root(builder.tree, &builder.file_data.file_text) {
+    let mut cursor = match Cursor::new_at_root(builder.tree, builder.file_data) {
         Ok(cursor) => cursor,
         Err(file_span) => {
             builder